daemon: move some helpers to their use api_*.go files
now that we have smaller api_*.go files some single use helpers
can live directly in them

move helpers from snap.go used by api_apps.go and api_find.go
